#6 Shotgun Surgery smell detector implemented and tested

+++ test/br/pucrio/opus/smells/tests/metrics/WMCTest.java
@@ -38,6 +38,6 @@ public class WMCTest {
 	@Test
 	public void ccTest() throws IOException {
 		Double wmc = getWMC("CC.java");
-		Assert.assertEquals(new Double(94), wmc);
+		Assert.assertEquals(new Double(108), wmc);
 	}
 }
